io.ligolw: moved patch_ligotimegps context manager
needed also for parsing segments with numpy types
